- Went on with GnucashGenerJob(Impl) and GnucashJobInvoice(Impl):    o Getting as much symmetric with applicable attrib. / methods      from GnucashCustomer(Invoice) and GnucashVendor(Invoice) as possible.    o Cleaned up logic of a job invoice's "owner" (direct vs. indirect):      As opposed to old approach (where e.t. implicitly went through jobs)      and the first generalized (and not so good) approach (variants of      get(Un)PaidInvoices_xyz-methods on the generic-invoice level),      we now have much more consistent and symmetric code.      Only for job invoices, we now have the ReadVariant.DIRECT/VIA_JOB      argument. For e.t. else, this does not make sense.    o Test cases now completely working again (as far as they go). However,      there is still work to do.  - As always, some code cleaning

+++ gnucash-api/src/test/java/org/gnucash/write/impl/TestGnucashWritableCustomerImpl.java
@@ -100,7 +100,8 @@ public class TestGnucashWritableCustomerImpl
       gcshInFile.writeFile(outFile);
       
       // copy file
-      FileUtils.delete(outFileGlob);
+      if ( outFileGlob.exists() )
+	  FileUtils.delete(outFileGlob);
       FileUtils.copyFile(outFile, outFileGlob);
   }
 
@@ -184,7 +185,8 @@ public class TestGnucashWritableCustomerImpl
       gcshInFile.writeFile(outFile);
       
       // copy file
-      FileUtils.delete(outFileGlob);
+      if ( outFileGlob.exists() )
+	  FileUtils.delete(outFileGlob);
       FileUtils.copyFile(outFile, outFileGlob);
   }
